WIP: applying theme from the URL
This removes all the magic from desktop URL rewriter. For logging out
we use native call, like on mobile, to be able to load the file. Before
it did not actually work and only loaded page after we've been failing
to load index.html/login.

TODO: updateCustomTheme() does not apply the theme yet